requirements.txt
IMDbPY==2021.4.18
PySimpleGUI==4.55.1
termcolor==1.1.0